Fix assess-robot-execute-kmacro, fix tests emacs<29
assess-robot-execute-kmacro was previously parsing the macro twice for
no obvious reason, which was causing breakage in Emacs-29.

The makefile now allows testing against various versions of Emacs
independently.

assess-test.el depends in a couple of places on the auto-mode-alist
binding which was broken by moving the fake elisp files to an "eld"
extension, for Emacs<29. We now set that explicitly.